added more test cases and code optimization
